make auth dir, move db endpoints to db dir
